+++ source/util/ArrayRef.h
@@ -27,10 +27,10 @@ public:
 
     ArrayRef(const T* begin, const T* end) : ptr(begin), len((uint32_t)(end - begin)) {}
 
-	ArrayRef(std::initializer_list<T> init) : ptr(init.begin()), len(init.size()) {}
+	ArrayRef(std::initializer_list<T> init) : ptr(init.begin()), len((uint32_t)init.size()) {}
 
 	template<typename Container>
-	ArrayRef(const Container& container) : ptr(container.data()), len(container.size()) {}
+	ArrayRef(const Container& container) : ptr(container.data()), len((uint32_t)container.size()) {}
 
     template<size_t N>
     ArrayRef(T (&array)[N]) : ptr(array), len(uint32_t(N)) {}
